pyproject.toml
[project]
name = "python-moal25"
version = "0.1.0"
description = "Python course assignments for moal25"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "go-task-bin>=3.44",
    "python-tester>=3",
    "ruff>=0.12",
]
